pyproject.toml
[tool.poetry]
name = "olive-python"
version = "1.0.0"
description = "BD103's Guide on Python Packaging"
authors = ["BD103 <dont@stalk.me>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
sphinx = "^4.0.2"
sphinx-better-theme = "^0.1.5"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
